service: move cluster server metrics into transport, drop gRPC alias
The "last command received" metric is no longer gRPC-specific, so move
ClusterServerMetrics/PerClusterServerMetrics out of the vestigial grpc
module and into the CTP transport module, and delete the now-empty
grpc.rs. Also drop the deprecated mz_grpc_server_last_command_received
alias; only mz_cluster_server_last_command_received is emitted now.
