+++ src/service/src/lib.rs
@@ -15,7 +15,6 @@
 
 pub mod boot;
 pub mod client;
-pub mod grpc;
 pub mod local;
 pub mod params;
 pub mod retry;
