+++ src/service/src/transport/metrics.rs
@@ -13,8 +13,13 @@ use std::io;
 use std::marker::PhantomData;
 use std::pin::Pin;
 use std::task::{Context, Poll};
+use std::time::UNIX_EPOCH;
 
+use mz_ore::metric;
+use mz_ore::metrics::{DeleteOnDropGauge, MetricsRegistry, UIntGaugeVec};
+use prometheus::core::AtomicU64;
 use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
+use tracing::error;
 
 /// A trait for types that observe connection metric events.
 pub trait Metrics<Out, In>: Clone + Send + 'static {
@@ -35,6 +40,63 @@ pub struct NoopMetrics;
 impl<Out, In> Metrics<Out, In> for NoopMetrics {
     fn bytes_sent(&mut self, _len: usize) {}
     fn bytes_received(&mut self, _len: usize) {}
+    fn message_sent(&mut self, _msg: &Out) {}
+    fn message_received(&mut self, _msg: &In) {}
+}
+
+/// Metrics for a cluster (CTP) server.
+pub struct ClusterServerMetrics {
+    last_command_received: UIntGaugeVec,
+}
+
+impl ClusterServerMetrics {
+    /// Registers the cluster server metrics into a `registry`.
+    pub fn register_with(registry: &MetricsRegistry) -> Self {
+        Self {
+            last_command_received: registry.register(metric!(
+                name: "mz_cluster_server_last_command_received",
+                help: "The time (in seconds since the Unix epoch) at which the server last \
+                       received data from the controller, including CTP keepalives. Used to \
+                       detect controller connections that are no longer reachable.",
+                var_labels: ["server_name"],
+            )),
+        }
+    }
+
+    /// Returns the metrics for the server with the given name.
+    pub fn for_server(&self, name: &'static str) -> PerClusterServerMetrics {
+        PerClusterServerMetrics {
+            last_command_received: self
+                .last_command_received
+                .get_delete_on_drop_metric(vec![name]),
+        }
+    }
+}
+
+/// Metrics for a single named cluster (CTP) server.
+#[derive(Clone, Debug)]
+pub struct PerClusterServerMetrics {
+    last_command_received: DeleteOnDropGauge<AtomicU64, Vec<&'static str>>,
+}
+
+impl<Out, In> Metrics<Out, In> for PerClusterServerMetrics {
+    fn bytes_sent(&mut self, _len: usize) {}
+
+    fn bytes_received(&mut self, _len: usize) {
+        // We bump the "last command received" timestamp on any inbound bytes, not just on fully
+        // decoded command messages (`message_received`). CTP exchanges keepalives on otherwise
+        // idle connections, and those show up here but never as messages (empty keepalive frames
+        // are dropped before decoding). Counting them as activity makes this metric a
+        // connection-liveness signal: it stays fresh as long as the controller connection is
+        // healthy and only goes stale once the connection is actually broken. Bumping it only on
+        // command messages instead produces false positives on healthy-but-idle clusters that
+        // don't receive a command for a while.
+        match UNIX_EPOCH.elapsed() {
+            Ok(ts) => self.last_command_received.set(ts.as_secs()),
+            Err(e) => error!("failed to get system time: {e}"),
+        }
+    }
+
     fn message_sent(&mut self, _msg: &Out) {}
     fn message_received(&mut self, _msg: &In) {}
 }
